+++ src/NeoVM/NeoVM.h
@@ -39,13 +39,13 @@ extern "C"
 	DllExport void __stdcall ExecutionEngine_Clean(ExecutionEngine* engine, uint32 iteration);
 	DllExport int32 __stdcall ExecutionEngine_LoadScript(ExecutionEngine* engine, byte* script, int32 scriptLength, int32 rvcount);
 	DllExport byte __stdcall ExecutionEngine_LoadCachedScript(ExecutionEngine* engine, int32 scriptIndex, int32 rvcount);
-	DllExport byte __stdcall ExecutionEngine_Execute(ExecutionEngine* engine, uint64 gas);
+	DllExport byte __stdcall ExecutionEngine_Execute(ExecutionEngine* engine, uint32 gas);
 	DllExport byte __stdcall ExecutionEngine_IncreaseGas(ExecutionEngine* engine, uint32 gas);
 	DllExport void __stdcall ExecutionEngine_StepInto(ExecutionEngine* engine);
 	DllExport void __stdcall ExecutionEngine_StepOver(ExecutionEngine* engine);
 	DllExport void __stdcall ExecutionEngine_StepOut(ExecutionEngine* engine);
 	DllExport byte __stdcall ExecutionEngine_GetState(ExecutionEngine* engine);
-	DllExport uint64 __stdcall ExecutionEngine_GetConsumedGas(ExecutionEngine* engine);
+	DllExport uint32 __stdcall ExecutionEngine_GetConsumedGas(ExecutionEngine* engine);
 	DllExport void __stdcall ExecutionEngine_AddLog(ExecutionEngine* engine, OnStepIntoCallback callback);
 
 	// StackItems
